download and bundle route updates
