+++ src/app/cube/library/library.component.ts
@@ -11,8 +11,7 @@ import { ChangelogService } from 'app/core/learning-object-module/changelog/chan
 import { LearningObjectService } from '../learning-object.service';
 import { trigger, style, group, transition, animate, query } from '@angular/animations';
 import { NavbarService } from 'app/core/client-module/navbar.service';
-import { BUNDLING_ROUTES } from 'app/core/learning-object-module/bundling/bundling.routes';
-import { BundlingService } from 'app/core/learning-object-module/bundling/bundling.service';
+
 @Component({
   selector: 'clark-library',
   templateUrl: './library.component.html',
@@ -84,8 +83,7 @@ export class LibraryComponent implements OnInit, OnDestroy {
     private changelogService: ChangelogService,
     private learningObjectService: LearningObjectService,
     private navbarService: NavbarService,
-    private notificationService: NotificationsService,
-    private bundlingService: BundlingService
+    private notificationService: NotificationsService
   ) { }
 
   @HostListener('window:resize', ['$event'])
@@ -234,15 +232,6 @@ export class LibraryComponent implements OnInit, OnDestroy {
       console.error(e);
     }
   }
-
-  // downloadObject(event: MouseEvent, object: LearningObject, index: number) {
-  //   event.stopPropagation();
-  //   this.currentIndex = index;
-  //   this.downloading[index] = true;
-  //   this.showDownloadModal = true;
-  //   this.bundlingService.downloadBundle(BUNDLING_ROUTES.DOWNLOAD_BUNDLE(object.id));
-  //   this.downloading[index] = false;
-  // }
 
   goToNotification(notification: any) {
     const parsedDetailsPath = notification.link.split('/');
